+++ packages/core-react/src/chains/index.ts
@@ -1,1 +1,1 @@
-export * from './holesky'
+export * from './holesky';