+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,20 @@
 # reef-knot
 
+## 1.9.0
+
+### Patch Changes
+
+- Updated dependencies [e2074ed]
+  - @reef-knot/connect-wallet-modal@1.7.0
+  - @reef-knot/wallets-list@1.5.0
+
+## 1.8.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/web3-react@1.5.0
+
 ## 1.7.4
 
 ### Patch Changes